companies/tradingagents/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
anthropic>=0.34
python-telegram-bot>=21.0
apscheduler>=3.10
python-dotenv>=1.0
httpx>=0.27
pytz>=2024.1
